+++ History.md
@@ -1,5 +1,7 @@
 ### 0.3.1
-* not bind the bindings not exist on model
+* not bind the bindings which not exist on model (by prop in model check)
+* refactor generateConfig api same as reactive constructor
+* allow filter with arguments
 
 ### 0.3.0
 * remove `data-format`, use filter instead